Remove basically redundant database calls and replace with just one
After analyzing it again, it seems that the calls getAutostartGameServers() and getAllGameServers() are very redundant. To reduce confusion, the caller should be the one to sort through the resulting list. This is further reinforced by the fact that the potential speed increase of the query based solution would be negligible since it's unlikely for the database to have a very large number of entries in the servers table.