Remove "enabled" field from DB
This was a feature that was supposed to be used in conjunction with the autostart field -- where you could have a server enabled but not autostart, so that you could manually turn on servers and they would autostop. This was an artifact of the previous system and given future plans to introduce an interactive system this feature seems redundant. This could come back in the future, but for now it is gone