Refactor Event and its descendants to require a TerminalMultiplexer to run commands
Remove public access for PATH_SCREEN and add MULTIPLEXER as a replacement. As a result, at the point of reading the properties file is when the correct multiplexer is instantiated. the rest of the program is meant to access the multiplexer through the properties file.

Note that screen is the only multiplexer supported at the moment.