consolidate game and moniker fields into one sessionName field
This is to determine the session name once per game instead of having each event do it. This will reduce errors in the future with potentially mismatched session names for the same server. I have removed getters for game and moniker because I don't believe they will be useful on their own but that may change later in the project.