Properly refactor ASCRepository for readability
I made a stupid mistake in a previous commit and forgot both spaces and
commas. Meaning that the queries were all something like
"SELECTsiddescription..." which is obviously wrong. So I opted for a
simpler and more readable String.format() solution.

Also switch to using `ON x = y` instead of `USING(x)` for safety.