Remove synchronization from ASCTerminal.readLine()
Previously, I synchronized ASCTerminal.readLine() in order to avoid the
situation where a thread could close the terminal just as execution in
another thread has just verified that the terminal is open. However,
this decision made it so that while one thread is reading, no thread
could write (because the print*() functions are synchronized too) --
which misses the point of a JLine terminal (async I/O).

The problem here is that readLine() is blocking, AND I want to
synchronize it so that a thread must wait for tasks to finish before
being able to close the terminal. Both these requirements make it so
that when a thread is reading a line no thread can close the terminal,
effectively softlocking the terminal. This problem existed in previous
commits, and I've just realized it.

Since the first requirement is more important, I've decided to remove
the synchronization on that method. Writes will still synchronize
amongst each other and with the close() method, meaning that log
messages will not get interrupted midway. However, the only issue now is
that there is a possibility we can call LineReader.readLine() when the
terminal is closed, which will raise a less verbose
IllegalStateException(). But this isn't really that big of a problem
because I already raise that same error when checking for closure.

Regardless, the bottom line is that while this situation can happen, it
should be very rare, and if it happens it's not really destructive as
the worst that can happen is that some user input is lost. The intent is
that the terminal is only closed AFTER the console is closed, so it
should be fine.