Implement event schedule loop
In other words, when the first batch of events have completed, it will
attempt to schedule again.

Also move shutdownHook to start() as it makes more sense

Note: I previously had a TODO there for replacing the shutdownHook with
a SignalHandler but as SO gracefully pointed out that class is in the
sun package which is not really documented nor really supported. The
shutdownHook on second thought seems like a better solution anyway.